Cache Field pipeline with extra pipes #144
Pipeline was being instantiated everytime serialize() was being run
this made it extremely hot code, in addition to this the extra pipes passed
to fields were being run inside of pipes.  Pipelines are now computed when the Field
is instantiated and all the extra pipes are found once.  This commit yields a further 10%
speed up.